+++ src/test/java/de/muenchen/dave/services/processzaehldaten/ProcessZaehldatenBelastungsplanServiceTest.java
@@ -15,11 +15,9 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-import de.muenchen.dave.TestUtils;
 import de.muenchen.dave.domain.Verkehrsbeziehung;
 import de.muenchen.dave.domain.Zeitintervall;
 import de.muenchen.dave.domain.dtos.OptionsDTO;
-import de.muenchen.dave.domain.dtos.laden.AbstractBelastungsplanDataDTO;
 import de.muenchen.dave.domain.dtos.laden.AbstractLadeBelastungsplanDTO;
 import de.muenchen.dave.domain.dtos.laden.BelastungsplanDataDTO;
 import de.muenchen.dave.domain.dtos.laden.BelastungsplanQjsDataDTO;
@@ -31,9 +29,7 @@ import de.muenchen.dave.domain.elasticsearch.Zaehlstelle;
 import de.muenchen.dave.domain.elasticsearch.ZaehlstelleRandomFactory;
 import de.muenchen.dave.domain.elasticsearch.Zaehlung;
 import de.muenchen.dave.domain.elasticsearch.ZaehlungRandomFactory;
-import de.muenchen.dave.domain.enums.FahrbewegungKreisverkehr;
 import de.muenchen.dave.domain.enums.Fahrzeug;
-import de.muenchen.dave.domain.enums.Himmelsrichtung;
 import de.muenchen.dave.domain.enums.TypeZeitintervall;
 import de.muenchen.dave.domain.enums.Zaehlart;
 import de.muenchen.dave.domain.enums.ZaehldatenIntervall;
@@ -48,16 +44,12 @@ import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
 import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.HashMap;
 import java.util.LinkedList;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 import java.util.Random;
 import java.util.Set;
 import java.util.UUID;
-import org.apache.commons.lang3.ArrayUtils;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -85,7 +77,7 @@ public class ProcessZaehldatenBelastungsplanServiceTest {
 
     @BeforeEach
     public void beforeEach() throws IllegalAccessException {
-        service = new ProcessZaehldatenBelastungsplanService(zeitintervallRepository, zaehlstelleIndex, ladeZaehldatenService);
+        service = new ProcessZaehldatenBelastungsplanService(zeitintervallRepository, zaehlstelleIndex, ladeZaehldatenService, new BelastungsplanDataServiceFactory(new BelastungsplanDataDefaultService(), new BelastungsplanDataQjsService(), new BelastungsplanDataFjsService()));
         Mockito.reset(zeitintervallRepository, zaehlstelleIndex, ladeZaehldatenService);
     }
 
@@ -306,156 +298,6 @@ public class ProcessZaehldatenBelastungsplanServiceTest {
     }
 
     @Test
-    public void testIsKreisverkehr() {
-        final Verkehrsbeziehung verkehrsbeziehung = new Verkehrsbeziehung();
-
-        verkehrsbeziehung.setVon(1);
-        verkehrsbeziehung.setNach(null);
-        verkehrsbeziehung.setFahrbewegungKreisverkehr(FahrbewegungKreisverkehr.HINEIN);
-        boolean result = TestUtils.privateStaticMethodCall(
-                "isKreisverkehr",
-                ProcessZaehldatenBelastungsplanService.class,
-                ArrayUtils.toArray(Verkehrsbeziehung.class),
-                ArrayUtils.toArray(verkehrsbeziehung),
-                Boolean.class);
-        assertThat(result, is(true));
-
-        verkehrsbeziehung.setVon(1);
-        verkehrsbeziehung.setNach(null);
-        verkehrsbeziehung.setFahrbewegungKreisverkehr(null);
-        result = TestUtils.privateStaticMethodCall(
-                "isKreisverkehr",
-                ProcessZaehldatenBelastungsplanService.class,
-                ArrayUtils.toArray(Verkehrsbeziehung.class),
-                ArrayUtils.toArray(verkehrsbeziehung),
-                Boolean.class);
-        assertThat(result, is(false));
-
-        verkehrsbeziehung.setVon(1);
-        verkehrsbeziehung.setNach(2);
-        verkehrsbeziehung.setFahrbewegungKreisverkehr(null);
-        result = TestUtils.privateStaticMethodCall(
-                "isKreisverkehr",
-                ProcessZaehldatenBelastungsplanService.class,
-                ArrayUtils.toArray(Verkehrsbeziehung.class),
-                ArrayUtils.toArray(verkehrsbeziehung),
-                Boolean.class);
-        assertThat(result, is(false));
-
-        verkehrsbeziehung.setVon(null);
-        verkehrsbeziehung.setNach(null);
-        verkehrsbeziehung.setFahrbewegungKreisverkehr(null);
-        result = TestUtils.privateStaticMethodCall(
-                "isKreisverkehr",
-                ProcessZaehldatenBelastungsplanService.class,
-                ArrayUtils.toArray(Verkehrsbeziehung.class),
-                ArrayUtils.toArray(verkehrsbeziehung),
-                Boolean.class);
-        assertThat(result, is(false));
-    }
-
-    @Test
-    public void testGetBelastungsplanData() {
-        final Map<Verkehrsbeziehung, ProcessZaehldatenBelastungsplanService.TupelTageswertZaehldatum> zaehldatenJeVerkehrsbeziehung = new HashMap<>();
-        Verkehrsbeziehung verkehrsbeziehung = new Verkehrsbeziehung();
-        verkehrsbeziehung.setVon(2);
-        verkehrsbeziehung.setNach(3);
-
-        LadeZaehldatumDTO ladeZaehldatum = new LadeZaehldatumDTO();
-        ladeZaehldatum.setPkw(1);
-        ladeZaehldatum.setLkw(2);
-        ladeZaehldatum.setLastzuege(3);
-        ladeZaehldatum.setBusse(4);
-        ladeZaehldatum.setKraftraeder(5);
-        ladeZaehldatum.setFahrradfahrer(6);
-        ladeZaehldatum.setFussgaenger(7);
-        ladeZaehldatum.setPkwEinheiten(100);
-
-        zaehldatenJeVerkehrsbeziehung.put(verkehrsbeziehung, new ProcessZaehldatenBelastungsplanService.TupelTageswertZaehldatum(false, ladeZaehldatum));
-
-        verkehrsbeziehung = new Verkehrsbeziehung();
-        verkehrsbeziehung.setVon(5);
-        verkehrsbeziehung.setNach(2);
-
-        ladeZaehldatum = new LadeZaehldatumDTO();
-        ladeZaehldatum.setPkw(10);
-        ladeZaehldatum.setLkw(20);
-        ladeZaehldatum.setLastzuege(30);
-        ladeZaehldatum.setBusse(40);
-        ladeZaehldatum.setKraftraeder(50);
-        ladeZaehldatum.setFahrradfahrer(60);
-        ladeZaehldatum.setFussgaenger(70);
-        ladeZaehldatum.setPkwEinheiten(1000);
-
-        zaehldatenJeVerkehrsbeziehung.put(verkehrsbeziehung, new ProcessZaehldatenBelastungsplanService.TupelTageswertZaehldatum(false, ladeZaehldatum));
-
-        final Zaehlung zaehlung = new Zaehlung();
-        zaehlung.setKategorien(Arrays.asList(Fahrzeug.KFZ, Fahrzeug.SV, Fahrzeug.GV, Fahrzeug.SV_P, Fahrzeug.GV_P, Fahrzeug.RAD, Fahrzeug.FUSS));
-        zaehlung.setZaehlart(Zaehlart.N.toString());
-        zaehlung.setKreisverkehr(false);
-
-        final Map<Fahrzeug, AbstractBelastungsplanDataDTO> belastungsplanData = new ProcessZaehldatenBelastungsplanService(null, null, null)
-                .getBelastungsplanData(zaehldatenJeVerkehrsbeziehung, zaehlung);
-
-        assertThat(((BelastungsplanDataDTO) belastungsplanData.get(Fahrzeug.KFZ)).getValues()[1][2], is(BigDecimal.valueOf(15)));
-        assertThat(((BelastungsplanDataDTO) belastungsplanData.get(Fahrzeug.SV)).getValues()[1][2], is(BigDecimal.valueOf(9)));
-        assertThat(((BelastungsplanDataDTO) belastungsplanData.get(Fahrzeug.GV)).getValues()[1][2], is(BigDecimal.valueOf(5)));
-        assertThat(((BelastungsplanDataDTO) belastungsplanData.get(Fahrzeug.RAD)).getValues()[1][2], is(BigDecimal.valueOf(6)));
-        assertThat(((BelastungsplanDataDTO) belastungsplanData.get(Fahrzeug.FUSS)).getValues()[1][2], is(BigDecimal.valueOf(7)));
-
-        assertThat(((BelastungsplanDataDTO) belastungsplanData.get(Fahrzeug.KFZ)).getValues()[4][1], is(BigDecimal.valueOf(150)));
-        assertThat(((BelastungsplanDataDTO) belastungsplanData.get(Fahrzeug.SV)).getValues()[4][1], is(BigDecimal.valueOf(90)));
-        assertThat(((BelastungsplanDataDTO) belastungsplanData.get(Fahrzeug.GV)).getValues()[4][1], is(BigDecimal.valueOf(50)));
-        assertThat(((BelastungsplanDataDTO) belastungsplanData.get(Fahrzeug.RAD)).getValues()[4][1], is(BigDecimal.valueOf(60)));
-        assertThat(((BelastungsplanDataDTO) belastungsplanData.get(Fahrzeug.FUSS)).getValues()[4][1], is(BigDecimal.valueOf(70)));
-
-    }
-
-    @Test
-    public void testGetBelastungsplanQjsData() {
-        final Map<Verkehrsbeziehung, ProcessZaehldatenBelastungsplanService.TupelTageswertZaehldatum> zaehldatenJeVerkehrsbeziehung = new HashMap<>();
-
-        Verkehrsbeziehung verkehrsbeziehung = new Verkehrsbeziehung();
-        verkehrsbeziehung.setVon(2);
-        verkehrsbeziehung.setNach(4);
-        verkehrsbeziehung.setStrassenseite(Himmelsrichtung.N);
-
-        LadeZaehldatumDTO ladeZaehldatum = new LadeZaehldatumDTO();
-        ladeZaehldatum.setFahrradfahrer(6);
-        ladeZaehldatum.setFussgaenger(7);
-
-        zaehldatenJeVerkehrsbeziehung.put(verkehrsbeziehung, new ProcessZaehldatenBelastungsplanService.TupelTageswertZaehldatum(false, ladeZaehldatum));
-
-        verkehrsbeziehung = new Verkehrsbeziehung();
-        verkehrsbeziehung.setVon(4);
-        verkehrsbeziehung.setNach(2);
-        verkehrsbeziehung.setStrassenseite(Himmelsrichtung.N);
-
-        ladeZaehldatum = new LadeZaehldatumDTO();
-        ladeZaehldatum.setFahrradfahrer(60);
-        ladeZaehldatum.setFussgaenger(70);
-
-        zaehldatenJeVerkehrsbeziehung.put(verkehrsbeziehung, new ProcessZaehldatenBelastungsplanService.TupelTageswertZaehldatum(false, ladeZaehldatum));
-
-        final Zaehlung zaehlung = new Zaehlung();
-        zaehlung.setKategorien(Arrays.asList(Fahrzeug.RAD, Fahrzeug.FUSS));
-        zaehlung.setZaehlart(Zaehlart.QJS.toString());
-        zaehlung.setKreisverkehr(false);
-
-        final Map<Fahrzeug, AbstractBelastungsplanDataDTO> belastungsplanData = new ProcessZaehldatenBelastungsplanService(null, null, null)
-                .getBelastungsplanQjsData(zaehldatenJeVerkehrsbeziehung, zaehlung);
-
-        List<BelastungsplanQjsDataDTO.VerkehrsbeziehungValue> valuesFuss = ((BelastungsplanQjsDataDTO) belastungsplanData.get(Fahrzeug.FUSS))
-                .getValuesVerkehrsbeziehungen();
-        assertVerkehrsbeziehung(valuesFuss, 2, 4, Himmelsrichtung.N, 7);
-        assertVerkehrsbeziehung(valuesFuss, 4, 2, Himmelsrichtung.N, 70);
-        List<BelastungsplanQjsDataDTO.VerkehrsbeziehungValue> valuesRad = ((BelastungsplanQjsDataDTO) belastungsplanData.get(Fahrzeug.RAD))
-                .getValuesVerkehrsbeziehungen();
-        assertVerkehrsbeziehung(valuesRad, 2, 4, Himmelsrichtung.N, 6);
-        assertVerkehrsbeziehung(valuesRad, 4, 2, Himmelsrichtung.N, 60);
-    }
-
-    @Test
     void testCalculateDifferenzdatenDTO() {
         final LadeBelastungsplanDTO dto1 = new LadeBelastungsplanDTO();
         BelastungsplanDataDTO belastungsplanData = new BelastungsplanDataDTO();
@@ -534,7 +376,7 @@ public class ProcessZaehldatenBelastungsplanServiceTest {
 
     @Test
     public void testRoundToNearestIfRoundingIsChoosen() {
-        final Integer nearestValueToRound = 100;
+        final int nearestValueToRound = 100;
         final OptionsDTO options = new OptionsDTO();
         options.setWerteHundertRunden(false);
         LadeZaehldatumDTO ladeZaehldatumDTO = new LadeZaehldatumDTO();
@@ -567,8 +409,8 @@ public class ProcessZaehldatenBelastungsplanServiceTest {
 
     @Test
     public void testRoundIfNotNullOrZero() {
-        final Integer nearestValueToRound = 100;
-        Integer valueToRoundInt = 49;
+        final int nearestValueToRound = 100;
+        int valueToRoundInt = 49;
         Integer resultInt = ProcessZaehldatenBelastungsplanService.roundIfNotNullOrZero(valueToRoundInt, nearestValueToRound);
         assertThat(resultInt, is(0));
 
@@ -739,15 +581,6 @@ public class ProcessZaehldatenBelastungsplanServiceTest {
         if (fahrzeuge.contains(Fahrzeug.FUSS))
             options.setFussverkehr(true);
         return options;
-    }
-
-    private void assertVerkehrsbeziehung(List<BelastungsplanQjsDataDTO.VerkehrsbeziehungValue> values, int von, int nach, Himmelsrichtung strassenseite,
-            int expectedValue) {
-        assertThat(values
-                .stream()
-                .filter(vb -> vb.getVon() == von && vb.getNach() == nach && vb.getStrassenseite().equals(strassenseite))
-                .map(BelastungsplanQjsDataDTO.VerkehrsbeziehungValue::getValue).findFirst().orElse(BigDecimal.ZERO),
-                is(BigDecimal.valueOf(expectedValue)));
     }
 
 }
